Fixed Issue #5
Key error caused by get_accounts referencing configuration file in some cases before the configuration file is written. Fix was to alter the function to expect an access token (the only thing used from the conf file) as a parameter instead of link_name (which was only used to find the right item in the conf file).

Also found and fixed another issue when an institution has no routing numbers found by Plaid.